% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm_generator.R
\name{bm_generator}
\alias{bm_generator}
\title{Generate matrix of coefficients for survival time generation}
\usage{
bm_generator(
  stime,
  num_bvar,
  num_tvar,
  bcov_names,
  tcov_names,
  treat_beta,
  violate,
  n
)
}
\arguments{
\item{stime}{Follow-up time for coefficient matrix}

\item{num_bvar}{Number of baseline coefficients}

\item{num_tvar}{Number of time-varying coefficients}

\item{bcov_names}{Names of baseline variables}

\item{tcov_names}{Names of time-varying variables}

\item{treat_beta}{Treatment coefficient}

\item{violate}{Tag for violation of estimation assumptions}

\item{n}{Number of patients}
}
\value{
A matrix of dimension [n*stime]x[num_tvar + num_bvar + 2]
}
\description{
Generate matrix of coefficients for survival time generation
}
