% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_generator.R
\name{.cc_generate}
\alias{.cc_generate}
\title{Generate time}
\usage{
.cc_generate(numb, numt)
}
\arguments{
\item{numb}{Number of baseline covariates}

\item{numt}{Number of time-varying covariates}
}
\value{
A list of 2 data.frames named 'baseline' and 'varying', representing baseline and time varying coefficients
for the generation of survival times via a cox-like hazard model.
}
\description{
Generate time
}
