% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_generator.R
\name{fd_generator}
\alias{fd_generator}
\title{Full (unadjusted) dataframe generator}
\usage{
fd_generator(bcov, id_trt, n, num_bvar, num_tvar, prop_trt, stime)
}
\arguments{
\item{bcov}{Data.frame of baseline variables, 1 per patient per follow-up time. Thus, dimensions are
[stime*n]x[num_bvar]}

\item{id_trt}{Patient ids of patients assigned to treatment}

\item{n}{Number of patients}

\item{num_bvar}{Number of baseline variables in returned data.frame}

\item{num_tvar}{Number of time-varying covariates in returned data.frame.}

\item{prop_trt}{Proportion of \code{n} that is assigned to treatment}

\item{stime}{Number of follow-up times}
}
\value{
A data.frame with variables 'ids', 'arm', 'switch', 'time', 'treat', 'b1', 'b2', 'b3',
'M', 'v1', 'v2', "Mtime'
}
\description{
Full (unadjusted) dataframe generator
}
