% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bcov.R
\name{gen_bcov}
\alias{gen_bcov}
\title{Generate a matrix of baseline covariates}
\usage{
gen_bcov(num_bvar, diags, middle, stime, n)
}
\arguments{
\item{n}{}
}
\value{

}
\description{
Generate a matrix of baseline covariates
}
