% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haz_func.R
\name{haz_func}
\alias{haz_func}
\title{Define a default hazard function}
\usage{
haz_func(t, x, betas, b_haz, ncov = 0, ...)
}
\arguments{
\item{...}{}
}
\value{
\code{haz_func()} returns a hazard at time \code{t}, given covariates \code{x},
baseline hazard \code{b_haz}, and coefficients \code{betas}
}
\description{
Define a default hazard function
}
