% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simswitchloop_pack.R
\name{simswitch}
\alias{simswitch}
\title{Simulate TS data and apply treatment-effect-estimating methods}
\usage{
simswitch(
  add_tvar = 0,
  b_allowance = 0.1,
  b_haz,
  b_mag = 0.5,
  b_scale,
  b_shape,
  bcov,
  beta.mat,
  cens_flag = "Random",
  covar_coef,
  dep = dep_func,
  haz = haz_func,
  hide_tvar = 0,
  ipcw_exclude = TRUE,
  m_allowance = 0.1,
  m_inflation = 2,
  m_fidelity = 0.2,
  m_hard = TRUE,
  m_haz,
  m_mag = 2,
  m_scale,
  m_shape,
  n = 400,
  num_bvar = 3,
  num_cores = 0,
  num_tvar = 3,
  para = FALSE,
  prop_cens,
  prop_cens_allowance = 0.1,
  prop_cont_event,
  prop_switch = 0.5,
  prop_trt = 0.5,
  prop_trt_event,
  recens = TRUE,
  reps = 1000,
  rerun_lim = 200,
  s_allowance = 0.1,
  s_haz = weihaz(1:stime, s_shape, s_scale),
  s_mag = 0.5,
  s_scale = 0.7 * stime,
  s_shape = 2,
  seed,
  stime = 100,
  switch_coef = c(runif(1, 1.5, 2), runif(num_bvar - 1, 0.1, 0.3), runif(num_tvar, 0.2,
    0.5)),
  t_allowance = 0.1,
  t_mag = 0.5,
  treat_beta = -1,
  tse_dist = "loglogistic",
  unfix = as.character(c()),
  verbose = 2,
  violate = "None"
)
}
\arguments{
\item{add_tvar}{Number of covariates to add which have no effect on overall survival, but which are nonetheless included in IPCW, TSE and RPSFTM models (non-negative integer).}

\item{b_haz}{Basic hazard function determining survival. In a simulation with no covariates and 0 treatment effect, this is the universal hazard function.
Default is the hazard associated with a weibull distribution with \emph{shape} = 1 and \emph{scale} = stime. A user-defined entry must be vector or array of length \emph{stime} representing the hazard up to and inculding each time point. For specifying a parametric basic hazard function, see arguments \emph{b_scale} and \emph{b_shape}.}

\item{b_mag}{The intial magnitude by which the hazard function \emph{b_haz} is adjusted if the desired number of events is not observed. Default is 0.5 (and the default value is usually sufficient). Must be postive and numeric.}

\item{covar_coef}{This is a new comment}

\item{hide_tvar}{add option for hiding relevent covariates from switching modeling}

\item{n}{Number of participants in simulated trials.}

\item{num_bvar}{Number of basline covariates to include.}

\item{num_tvar}{Number of time-varying covariates to include.}

\item{prop_cens}{Proportion of patients}

\item{prop_switch}{set proportion of control participants to switch}

\item{prop_trt_event}{set proportion of participants on experimental treatment}

\item{rerun_lim}{set while loop limit}

\item{violate}{}
}
\value{

}
\description{
Simulate TS data and apply treatment-effect-estimating methods
}
