% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simswitchloop_pack.R
\docType{class}
\name{SwitchSimulation-class}
\alias{SwitchSimulation-class}
\title{The simswitch result class}
\value{

}
\description{
The simswitch result class
}
\section{Slots}{

\describe{
\item{\code{unbiased}}{numeric.}

\item{\code{unbiased_plot}}{ggsurvplot.}

\item{\code{itt}}{numeric.}

\item{\code{itt_plot}}{ggsurvplot.}

\item{\code{pp}}{numeric.}

\item{\code{pp_plot}}{ggsurvplot.}

\item{\code{ipcw}}{numeric.}

\item{\code{ipcw_plot}}{ggsurvplotOrNULL.}

\item{\code{rpsft}}{numeric.}

\item{\code{rpsft_plot}}{ggsurvplot.}

\item{\code{tse}}{numeric.}

\item{\code{tse_plot}}{ggsurvplot.}

\item{\code{compar_plot}}{gg.}

\item{\code{bias_plot}}{gg.}

\item{\code{params}}{list.}
}}

