% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simswitch.R
\name{simswitch}
\alias{simswitch}
\title{Simulate TS data and apply treatment-effect-estimating methods}
\usage{
simswitch(
  add_tvar,
  b_allowance,
  b_haz,
  b_mag,
  b_scale,
  b_shape,
  bcov,
  beta.mat,
  bootrep,
  cens_flag,
  covar_coef,
  dep_func,
  haz,
  hide_tvar,
  ipcw_robust,
  m_allowance,
  m_inflation,
  m_fidelity,
  m_hard,
  m_haz,
  m_mag,
  m_scale,
  m_shape,
  n,
  num_bvar,
  num_tvar,
  prop_cens,
  prop_cens_allowance,
  prop_cont_event,
  prop_switch,
  prop_trt,
  prop_trt_event,
  recens,
  rerun_lim,
  s_allowance,
  s_haz,
  s_mag,
  s_scale,
  s_shape,
  stime,
  switch_coef,
  t_allowance,
  t_mag,
  treat_hr,
  tse_dist,
  verbose,
  violate
)
}
\arguments{
\item{violate}{}
}
\value{

}
\description{
Simulate TS data and apply treatment-effect-estimating methods
}
